kodosumi-status/requirements.txt
asyncpg==0.29.0
httpx==0.25.0
python-dotenv==1.0.0